# internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package database

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Scroll struct {
	ID        string
	JarID     string
	Title     pgtype.Text
	Format    pgtype.Text
	Uploaded  bool
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type Scrolljar struct {
	ID           string
	Name         pgtype.Text
	UserID       pgtype.Int8
	Access       int16
	PasswordHash pgtype.Text
	Tags         []string
	ExpiresAt    pgtype.Timestamptz
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}

type Token struct {
	TokenHash []byte
	UserID    int64
	ExpiresAt pgtype.Timestamptz
	Scope     string
}

type UserAccount struct {
	ID           int64
	Username     string
	Email        string
	PasswordHash string
	Activated    bool
	CreatedAt    pgtype.Timestamptz
	UpdatedAt    pgtype.Timestamptz
}