# internal/spec/api_spec.yaml
openapi: 3.0.3
info:
  title: Scrolljar API
  description: |-
    API interface to work with Scrolljar
  version: 1.0.0

servers:
  - url: https://localhost:8008/v1

tags:
  - name: Misc
    description: API server health and other endpoints
  - name: Jar
    description: Operations for reading, creating and deleting jars.
  - name: Scroll
    description: Operations for CRUD operation on scrolls
  - name: User
    description: Operations about user (registration, activation and auth)
  - name: Token
    description: Operations to generate new user token

paths:
  /ping:
    get:
      tags: [Misc]
      summary: Ping to get health of server.
      operationId: ping
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ping'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /jar:
    post:
      tags: [Jar]
      summary: Route to create a new Jar
      operationId: createJar
      requestBody:
        $ref: '#/components/requestBodies/CreateJarInput'
      responses:
        '200':
          $ref: '#/components/responses/CreateJarOutput'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
  /jar/{id}:
    get:
      tags: [Jar]
      summary: Route to get a jar information
      operationId: getJar
      parameters:
        - $ref: '#/components/parameters/JarId'
        - name: X-Paste-Password
          in: header
          required: false
          schema:
            type: string
          description: Optional password for password protected jar
      responses:
        '200':
          $ref: '#/components/responses/Jar'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

    delete:
      tags: [Jar]
      summary: Route to delete a Jar. Deleting a Jar will all the scrolls within it.
      operationId: deleteJar
      parameters:
        - $ref: '#/components/parameters/JarId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulMessage'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []

  /jar/{id}/scrolls:
    get:
      tags: [Scroll]
      summary: Route to get all scrolls of a Jar
      operationId: getJarScrolls
      parameters:
        - $ref: '#/components/parameters/JarId'
      responses:
        '200':
          $ref: '#/components/responses/ScrollCollection'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /scroll/{id}:
    post:
      tags: [Scroll]
      summary: Route to create a new Scroll
      operationId: createScroll
      parameters:
        - $ref: '#/components/parameters/JarId'
      requestBody:
        $ref: '#/components/requestBodies/CreateScrollInput'
      responses:
        '200':
          $ref: '#/components/responses/CreateScrollOutput'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []
    
    get:
      tags: [Scroll]
      summary: Route to get a scroll information
      operationId: getScroll
      parameters:
        - $ref: '#/components/parameters/ScrollId'
        - name: X-Paste-Password
          in: header
          required: false
          schema:
            type: string
          description: Optional password for password protected jar
      responses:
        '200':
          $ref: '#/components/responses/ScrollFetch'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

    delete:
      tags: [Scroll]
      summary: Route to delete a scroll
      operationId: deleteScroll
      parameters:
        - $ref: '#/components/parameters/ScrollId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulMessage'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []

    patch:
      tags: [Scroll]
      summary: Route to update a scroll
      operationId: patchScroll
      parameters:
        - $ref: '#/components/parameters/ScrollId'
      requestBody:
        $ref: '#/components/requestBodies/ScrollPatchInput'
      responses:
        '200':
          $ref: '#/components/responses/CreateScrollOutput'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []

  /upload:
    put:
      tags: [Scroll]
      summary: Route to upload the scroll content
      operationId: uploadScroll
      parameters:
        - name: X-Upload-Token
          in: header
          required: true
          schema:
            type: string
          description: Upload token to upload the content 
      responses:
        '200':
          $ref: '#/components/responses/ScrollFetch'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /user:
    get:
      tags: [User]
      summary: Route to get information about the user and the jars created by them
      operationId: getUser
      responses:
        '200':
          $ref: '#/components/responses/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []

  /user/jars:
    get:
      tags: [Jar]
      summary: Route to get list of jars creatd by the user
      operationId: getUserJars
      responses:
        '200':
          $ref: '#/components/responses/JarCollection'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'
      security:
        - BearerAuth: []

  /user/register:
    post:
      tags: [User]
      summary: Route to create a new User
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/RegistrationInput'
      responses:
        '200':
          $ref: '#/components/responses/User'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /user/activate:
    put:
      tags: [User]
      summary: Route to activate a new User
      operationId: activateUser
      requestBody:
        $ref: '#/components/requestBodies/ActivationInput'
      responses:
        '200':
          $ref: '#/components/responses/SuccessfulMessage'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /user/auth:
    post:
      tags: [User]
      summary: Route to authenticate and get an access token
      operationId: authUser
      requestBody:
        $ref: '#/components/requestBodies/LoginInput'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

  /token/activation:
    post:
      tags: [Token]
      summary: Route to get a activation token of a user
      operationId: createActivationToken
      requestBody:
        $ref: '#/components/requestBodies/LoginInput'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        default:
          $ref: '#/components/responses/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  parameters:
    JarId:
      name: id
      in: path 
      required: true
      schema:
        type: string

    ScrollId:
      name: id
      in: path 
      required: true
      schema:
        type: string

  requestBodies:
    CreateJarInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateJarInput'

    CreateScrollInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateScrollInput'
    
    ScrollPatchInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ScrollPatchInput'

    RegistrationInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegistrationInput'

    ActivationInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ActivationInput'

    LoginInput:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginInput'

  responses:
    RateLimitExceeded:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Item not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Unauthorized user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

    SuccessfulMessage:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'

    Scroll:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Scroll'
    
    ScrollFetch:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ScrollFetch'
    
    Jar:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Jar'
    
    CreateJarOutput:
      description: Operation Successful
      content:
        application/json: 
          schema:
            $ref: '#/components/schemas/CreateJarOutput'
      
    CreateScrollOutput:
      description: Operation Successful
      content:
        application/json: 
          schema:
            $ref: '#/components/schemas/CreateScrollOutput'

    User:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
    
    JarCollection:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JarCollection'
    
    ScrollCollection:
      description: Operation Successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ScrollCollection'

  schemas:
    Ping:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
        environment:
          type: string
        version:
          type: string
        uptime:
          type: string
      required: [status, environment, version, uptime]

    Message:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
    
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
    
    FieldError:
      type: object
      required: [field, msg]
      additionalProperties: false
      properties:
        field:
          type: array
          description: array showing nested location of errorous field
          example: [scrolls, content]
          items:
            type: string
        msg:
          description: error message
          type: string

    ValidationError:
      type: object
      additionalProperties: false
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    AuthTokens:
      type: object
      additionalProperties: false
      required: [authorization, refresh]
      properties:
        authorization:
          $ref: '#/components/schemas/Token'
        refresh:
          $ref: '#/components/schemas/Token'

    JarAccess:
      type: integer
      enum: [0, 1]
      default: 0
      x-enum-varnames: [AccessPublic, AccessPrivate]

    Jar:
      type: object
      additionalProperties: false
      required: [id, access, tags, expires_at, created_at, uri]
      properties:
        id:
          type: string
        name:
          type: string
        access:
          $ref: '#/components/schemas/JarAccess'
        tags:
          type: array
          items:
            type: string
          x-go-type: pgtype.FlatArray[string]
          x-go-type-import:
            path: github.com/jackc/pgx/v5/pgtype
        expires_at:
          type: string
          format: date-time
          x-go-type: pgtype.Timestamptz
          x-go-type-import:
            path: github.com/jackc/pgx/v5/pgtype
        created_at:
          type: string
          format: date-time
          x-go-type: pgtype.Timestamptz
          x-go-type-import:
            path: github.com/jackc/pgx/v5/pgtype
        uri:
          type: string
          format: uri
      
    CreateJarOutput:
      type: object
      additionalProperties: false
      required: [jar, scrolls]
      properties:
        jar:
          $ref: "#/components/schemas/Jar"
        scrolls:
          type: array
          items:
            $ref: "#/components/schemas/CreateScrollOutput"
    
    CreateScrollOutput:
      type: object
      additionalProperties: false
      required: [upload_token, expiry]
      properties:
        scroll:
          $ref: "#/components/schemas/Scroll"
        upload_token:
          type: string

    JarCollection:
      type: array 
      items:
        $ref: '#/components/schemas/Jar'

    Scroll:
      type: object
      additionalProperties: false
      required: [id, jarid, created_at, uri]
      properties:
        id:
          type: string
        jarid:
          type: string
        title:
          type: string
        format:
          type: string
        created_at:
          type: string
          format: date-time
          x-go-type: pgtype.Timestamptz
          x-go-type-import:
            path: github.com/jackc/pgx/v5/pgtype
        uri:
          type: string
          format: uri
    
    ScrollFetch:
      type: object
      additionalProperties: false
      properties:
        scroll:
          $ref: '#/components/schemas/Scroll'
        fetch_url:
          type: string

    ScrollCollection:
      type: array 
      items:
        $ref: '#/components/schemas/Scroll'

    User:
      type: object
      additionalProperties: false
      required: [id, username, created_at]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        created_at:
          type: string
          format: date-time

    Token:
      type: object
      additionalProperties: false
      required: [token, expiry]
      properties:
        token:
          type: string
        expiry:
          type: string
          format: date-time

    CreateJarInput:
      type: object
      additionalProperties: false
      required: [name, scrolls]
      properties:
        name:
          type: string
        access:
          $ref: '#/components/schemas/JarAccess'
        password:
          type: string
          minLength: 1
        expiry:
          type: string
          x-go-type: ExpiryDuration
        tags:
          type: array
          items:
            type: string
        scrolls:
          type: array
          minItems: 1
          maxItems: 254
          items:
            $ref: '#/components/schemas/CreateScrollInput'

    CreateScrollInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        format:
          type: string
    
    ScrollPatchInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          x-go-type-skip-optional-pointer: false
        format:
          type: string
          x-go-type-skip-optional-pointer: false

    RegistrationInput:
      type: object
      additionalProperties: false
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 512
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72

    ActivationInput:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1

    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72